requirements.txt
streamlit>=1.28.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0
PyPDF2>=3.0.1
python-pptx>=0.6.21
python-docx>=0.8.11
pandas>=2.1.0
numpy>=1.24.0
google-generativeai>=0.8.5
langchain>=0.0.350
streamlit>=1.28.0